pyproject.toml
[project]
name = "dynaworld"
version = "0.1.0"
description = "Local TokenGS experiments"
requires-python = ">=3.11,<3.13"
dependencies = [
    "beartype",
    "gradio>=6.12.0",
    "moviepy>=2.2.1",
    "opencv-python",
    "packaging",
    "pillow",
    "requests",
    "roma",
    "taichi",
    "tensordict>=0.7.0",
    "torch",
    "torchvision",
    "tqdm",
    "wandb",
]

[tool.uv]
package = false
